MSVC warnings.
--HG--
branch : develop

+++ src/libs/vpropertyexplorer/plugins/Vector3d/vvector3dproperty.h
@@ -40,7 +40,7 @@ public:
 
     ~Vector3D() {}
 
-    float X, Y, Z;
+    double X, Y, Z;
 };
 /*
 }
